close drawer on navigation
